requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-countries==7.5.1
django-environ==0.10.0
django-language-field==0.0.3
django-multiselectfield==0.1.12
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
